lab2/pyproject.toml
[tool.poetry]
name = "mbi"
version = "0.1.0"
description = ""
authors = ["Piotr Hondra <piotr.hondra@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.81"

[tool.poetry.dev-dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
